+++ mayan/apps/linking/tests/test_views.py
@@ -260,3 +260,44 @@ class SmartLinkDocumentViewTestCase(
             count=5, response=response, status_code=200,
             text=self.test_document.label
         )
+
+    def test_document_resolved_smart_list_with_no_permission(self):
+        response = self._request_test_document_resolved_smart_link_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_resolved_smart_list_with_document_access(self):
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_resolved_smart_link_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_resolved_smart_list_with_smart_link_access(self):
+        self.grant_access(
+            obj=self.test_smart_links[0],
+            permission=permission_smart_link_view
+        )
+        self.grant_access(
+            obj=self.test_smart_links[1],
+            permission=permission_smart_link_view
+        )
+
+        response = self._request_test_document_resolved_smart_link_view()
+        self.assertEqual(response.status_code, 404)
+
+    def test_document_resolved_smart_list_with_full_access(self):
+        self.grant_access(
+            obj=self.test_smart_links[0],
+            permission=permission_smart_link_view
+        )
+        self.grant_access(
+            obj=self.test_smart_links[1],
+            permission=permission_smart_link_view
+        )
+        self.grant_access(
+            obj=self.test_document, permission=permission_document_view
+        )
+
+        response = self._request_test_document_resolved_smart_link_view()
+        self.assertEqual(response.status_code, 200)